requirement.txt
alabaster==0.7.8
alembic==0.8.6
Babel==2.3.4
blinker==1.4
click==6.6
docutils==0.12
enum==0.4.6
Flask==0.11.1
Flask-Cors==2.1.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Foundations==2.1.0
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.8
linecache2==1.0.0
Mako==1.0.4
Manager==2.0.5
MarkupSafe==0.23
Oncilla==0.1.0
pycparser==2.18
Pygments==2.1.3
PyJWT==0.2.1
PyMySQL==0.7.4
python-editor==1.0.1
pytz==2016.4
requests==2.3.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.4
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.13
traceback2==1.4.0
unittest2==1.1.0
uWSGI==2.0.15
Werkzeug==0.11.10
wheel==0.24.0
